+++ docs/changelog.md
@@ -21,6 +21,7 @@ Click on the version in the following list to see all changes since the previous
 * docs: add icons to introduction page
 * refactor: introduce service interface for the estimation field service
 * feat: add error logging capabilities to all possible error locations
+* docs: add documentation to ScrumPokerConstants class
 
 ### [4.2.0] - 2019-08-03